refactor : 객체 분리
