+++ src/main/kotlin/model/GameBoard.kt
@@ -2,22 +2,6 @@ package model
 
 import model.judge.OmokJudge
 
-sealed class AddStoneStatus {
-    data object IsExist : AddStoneStatus()
-
-    data object IsThreeThree : AddStoneStatus()
-
-    data object IsFourFour : AddStoneStatus()
-
-    data object IsWin : AddStoneStatus()
-
-    data object IsAble : AddStoneStatus()
-
-    data object IsOverFive : AddStoneStatus()
-
-    data object IsUnAblePosition : AddStoneStatus()
-}
-
 class GameBoard {
     private val _stones = mutableListOf<Stone>()
     val stones get() = _stones.toList()
